change uint32_t to uintptr_t where it violates precision losses

+++ modules/nrfx/drivers/include/nrfx_ppi.h
@@ -288,9 +288,9 @@ nrfx_err_t nrfx_ppi_group_disable(nrf_ppi_channel_group_t group);
  *
  * @retval Task address.
  */
-__STATIC_INLINE uint32_t nrfx_ppi_task_addr_get(nrf_ppi_task_t task)
+__STATIC_INLINE uintptr_t nrfx_ppi_task_addr_get(nrf_ppi_task_t task)
 {
-    return (uint32_t) nrf_ppi_task_address_get(task);
+    return (uintptr_t) nrf_ppi_task_address_get(task);
 }
 
 /**
@@ -300,9 +300,9 @@ __STATIC_INLINE uint32_t nrfx_ppi_task_addr_get(nrf_ppi_task_t task)
  *
  * @retval Task address.
  */
-__STATIC_INLINE uint32_t nrfx_ppi_task_addr_group_enable_get(nrf_ppi_channel_group_t group)
+__STATIC_INLINE uintptr_t nrfx_ppi_task_addr_group_enable_get(nrf_ppi_channel_group_t group)
 {
-    return (uint32_t) nrf_ppi_task_group_enable_address_get(group);
+    return (uintptr_t) nrf_ppi_task_group_enable_address_get(group);
 }
 
 /**
@@ -312,9 +312,9 @@ __STATIC_INLINE uint32_t nrfx_ppi_task_addr_group_enable_get(nrf_ppi_channel_gro
  *
  * @retval Task address.
  */
-__STATIC_INLINE uint32_t nrfx_ppi_task_addr_group_disable_get(nrf_ppi_channel_group_t group)
+__STATIC_INLINE uintptr_t nrfx_ppi_task_addr_group_disable_get(nrf_ppi_channel_group_t group)
 {
-    return (uint32_t) nrf_ppi_task_group_disable_address_get(group);
+    return (uintptr_t) nrf_ppi_task_group_disable_address_get(group);
 }
 
 /** @} */
